Move Rule into rule.go.
